[add]SaveDataLoaderをSaveDataLoaderStrategyへ改名
セーブデータの保存先を差し替える拡張点へ、役割を表すサフィックスを
付けた。中身（protected abstract メンバー）は変更していない。

- SaveDataLoaderStrategy / PlayerPrefsSaveDataLoaderStrategy を新設
- 旧型は新しい基底を継承した空の抽象クラスとして [Obsolete] で残す。
  旧型を継承した既存のローダーはそのまま動作する。削除は 3.0.0
- 内部の JsonUtilitySaveDataLoader / NewtonsoftSaveDataLoader も改名し、
  [MovedFrom] を付与

[MovedFrom] が必須な理由: SaveSystemConfig は [SerializeReference] で
{class, ns, asm} をアセットへ焼いている。属性が無いと既存の設定アセットで
ローダーが解決できず、利用側のローダー選択が失われる。

旧型名のままのアセットを実際に読み込ませ、
JsonUtilitySaveDataLoaderStrategy として解決されることを確認した。
アセットは検証中に書き換えていない。

利用側の移行も模擬し、旧基底を継承したローダーが CS0618 を出したうえで
コンパイルが通ることを確認した。

コンパイル エラー0・SymphonyFrameWork由来の警告0、EditMode 222件を
2回連続で成功、PlayMode 4件成功。
