+++ Editor/PackageInitializer.cs
@@ -95,11 +95,11 @@ namespace SymphonyFrameWork.Editor
         ///     Editor上の現在のConfigからセーブデータローダーを解決する。
         /// </summary>
         /// <returns> Configで選択されたローダー。未設定の場合は既定のローダー。 </returns>
-        private static SaveDataLoader ResolveSaveDataLoader()
+        private static SaveDataLoaderStrategy ResolveSaveDataLoader()
         {
             SaveSystemConfig config =
                 SymphonyConfigLocator.GetConfig<SaveSystemConfig>();
-            return config?.Loader ?? new JsonUtilitySaveDataLoader();
+            return config?.Loader ?? new JsonUtilitySaveDataLoaderStrategy();
         }
     }
 }
