+++ Runtime/System/SaveSystem/SaveDataLoaderStrategy.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ac6a2884ffeb4a69ae61ac2cf3e50dce
