+++ Runtime/System/SaveSystem/Template/PlayerPrefsSaveDataLoader.cs
@@ -1,45 +1,15 @@
 ﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
-using UnityEngine;
 
 namespace SymphonyFrameWork.System.SaveSystem
 {
     /// <summary>
-    ///     PlayerPrefsへのJSON入出力を提供するLoader基底クラスです。
+    ///     <see cref="PlayerPrefsSaveDataLoaderStrategy"/>へ置き換えられた非推奨の基底クラスです。
+    ///     この型を継承した既存のローダーは、新しい基底の派生としてそのまま動作します。
+    ///     メンバーを持たないため、実装は<see cref="PlayerPrefsSaveDataLoaderStrategy"/>が担います。
     /// </summary>
     [Serializable]
-    public abstract class PlayerPrefsSaveDataLoader : SaveDataLoader
+    [Obsolete("PlayerPrefsSaveDataLoaderStrategyを継承してください。3.0.0で削除します。", error: false)]
+    public abstract class PlayerPrefsSaveDataLoader : PlayerPrefsSaveDataLoaderStrategy
     {
-        /// <summary> PlayerPrefsに型固有のキーが存在するか確認する。 </summary>
-        protected override bool ExistsCore(Type dataType)
-        {
-            return PlayerPrefs.HasKey(GetKey(dataType));
-        }
-
-        /// <summary> PlayerPrefsから型固有のJSONを読み込む。 </summary>
-        protected override ValueTask<string> LoadJsonAsync(Type dataType, CancellationToken token)
-        {
-            return new ValueTask<string>(PlayerPrefs.GetString(GetKey(dataType)));
-        }
-
-        /// <summary> PlayerPrefsへ型固有のJSONを書き込む。 </summary>
-        protected override ValueTask SaveJsonAsync(Type dataType, string json, CancellationToken token)
-        {
-            PlayerPrefs.SetString(GetKey(dataType), json);
-            PlayerPrefs.Save();
-            return default;
-        }
-
-        /// <summary> PlayerPrefsから型固有の保存値を削除する。 </summary>
-        protected override ValueTask DeleteCoreAsync(Type dataType, CancellationToken token)
-        {
-            PlayerPrefs.DeleteKey(GetKey(dataType));
-            PlayerPrefs.Save();
-            return default;
-        }
-
-        /// <summary> 型の完全修飾名からPlayerPrefsキーを生成する。 </summary>
-        private static string GetKey(Type dataType) => dataType.FullName;
     }
 }
