+++ Runtime/System/SaveSystem/Template/PlayerPrefsSaveDataLoader.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: d307f08101b740978c90d14b799b926f
+guid: 5bd19fe784c429543aee3afaceac046e
