+++ Runtime/System/SaveSystem/Template/PlayerPrefsSaveDataLoaderStrategy.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d307f08101b740978c90d14b799b926f
