+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "symphonyframework",
-  "version": "2.19.0",
+  "version": "2.20.0",
   "displayName": "Symphony Framework",
   "description": "This is Symphony Framework",
   "unity": "6000.0",
